*From previous commit
Added disabled feature to show the user that the button can't be pressed until a score is selected. The opacity is lower to show it's not accessible. The cursor has been changed to 'not-allowed' to show it can't be done. One selected, the opacity increases showing the user the button can now be used.