Applied disabling to the submit button
Managing when the state of the button should change when the user selects a value to unlock the submit button. 
At the start, the button is declared as disabled and uses the class from CSS to set it in the styled state. 
Once picked a value, the class is removed and the button disabling is set to false.