Changed margins, changed centring and removed using pixels as unit
Removed margins from all tags to hide the scrollbar but this meant having to modify other components so their margins were corrected (it removed their margins as well).

Centring was in a div called viewport but this was put into the body by defining the height and width to be the whole of the viewport. This meant flex could be sued by defining the align-items and justify-content to be centred. 

Removed using pixels as unit so the page can be dynamic - except from border-radius which will need to stay the same on all devices.